refactor user entity and delete, softdelete methods for all entities

+++ Application/Features/Users/Commands/Delete/DeleteUserCommand.cs
@@ -1,7 +1,9 @@
 ﻿using Application.Features.Users.Constants;
 using Application.Repositories;
+using AutoMapper;
 using Core.Application.Pipelines.Authorization;
 using Core.Application.Pipelines.Logging;
+using Core.CrossCuttingConcerns.Exceptions.Types;
 using Domain.Entities;
 using MediatR;
 using static Application.Features.Users.Constants.UsersOperationClaims;
@@ -9,31 +11,35 @@ using static Application.Features.Users.Constants.UsersOperationClaims;
 
 namespace Application.Features.Users.Commands.Delete
 {
-    public class DeleteUserCommand : IRequest, ISecuredRequest, ILoggableRequest
+    public class DeleteUserCommand : IRequest<DeleteUserReponse>, ISecuredRequest, ILoggableRequest
     {
         public string[] RequiredRoles => new[] { Admin, UsersOperationClaims.Delete };
         public int Id { get; set; }
 
-        public class DeleteUserCommandHandler : IRequestHandler<DeleteUserCommand>
+        public class DeleteUserCommandHandler : IRequestHandler<DeleteUserCommand, DeleteUserReponse>
         {
             private readonly IUserRepository _userRepository;
+            private readonly IMapper _mapper;
 
-            public DeleteUserCommandHandler(IUserRepository userRepository)
+            public DeleteUserCommandHandler(IUserRepository userRepository, IMapper mapper)
             {
                 _userRepository = userRepository;
+                _mapper = mapper;
             }
 
-            public async Task Handle(DeleteUserCommand request, CancellationToken cancellationToken)
+            public async Task<DeleteUserReponse> Handle(DeleteUserCommand request, CancellationToken cancellationToken)
             {
                 
                 User? user = await _userRepository.GetAsync(i => i.Id == request.Id);
 
-                if (user == null)
+                if (user == null || user.IsDeleted == true)
                 {
-                    throw new ArgumentException("No such user found");
+                    throw new NotFoundException(UsersMessages.UserNotExists);
                 }
 
                 await _userRepository.DeleteAsync(user);
+                DeleteUserReponse response = _mapper.Map<DeleteUserReponse>(user);
+                return response;
 
             }
         }
